Changed retreat size from uint32_t to uint8_t. Retreat values would oscilate from 0 to 100, so it is a waste of bandwich spend 4 bytes storing that value. Removed unused params on event allow_retreat_change.
Also that values are saved in and loaded from savegames or netgames.

git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@4811 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/ai/defaultai.cc
@@ -2018,7 +2018,7 @@ bool DefaultAI::consider_attack(int32_t gametime) {
 	Building * target = ms; // dummy initialisation to silence the compiler
 	int32_t    chance = 0;
 	int32_t    attackers = 0;
-	int32_t    retreat = ms->owner().get_retreat_percentage();
+	int8_t     retreat = ms->owner().get_retreat_percentage();
 
 	// Search in a radius of the vision of the militarysite and collect
 	// information about immovables in the area
